requirements.txt
fastapi
uvicorn 
requests
faiss-cpu
ollama
python-docx